refactoring into DSP package
