+++ libs/automation/ns.ts
@@ -2,6 +2,7 @@
  * Automation, process control and robotic controllers
  */
 //% groups="['PID', `Filters', 'Behaviors']"
+//% color=0x909090 weight=80 icon="\uf277"
 namespace automation {
 
 }
