+++ libs/dsp/ns.ts
@@ -1,0 +1,8 @@
+/**
+ * Digital signal processing
+ */
+//% groups="['Filters']"
+//% color=0x909090 weight=80 icon="\uf277"
+namespace dsp {
+
+}
